added cgo for sqlite cross-compile

+++ core/internal/web/controllers/adminctrl/power-ctrl.go
@@ -2,6 +2,7 @@ package adminctrl
 
 import (
 	"core/internal/api"
+	"core/internal/utils/updates"
 	powerview "core/resources/views/admin/power"
 	"fmt"
 	"net/http"
@@ -22,13 +23,27 @@ func RebootPageCtrl(g *api.CoreGlobals) http.HandlerFunc {
 
 func RebootCtrl(g *api.CoreGlobals) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		// Check if sysupgrade file is ready
+		isSysupgrade := updates.IsSysupgradeReady()
+
 		go func() {
 			time.Sleep(3 * time.Second)
-			cmd.Exec("reboot", nil)
+			if isSysupgrade {
+				// Run sysupgrade command (this will flash firmware and reboot)
+				cmd.Exec("sysupgrade "+updates.GetSysupgradePath(), nil)
+			} else {
+				// Normal reboot
+				cmd.Exec("reboot", nil)
+			}
 		}()
 
 		w.Header().Set("Content-Type", "text/html")
-		msg := g.CoreAPI.Translate("info", "System is rebooting... Please wait a few minutes before reconnecting.")
+		var msg string
+		if isSysupgrade {
+			msg = g.CoreAPI.Translate("info", "System is upgrading firmware... This may take several minutes. Do not power off the device.")
+		} else {
+			msg = g.CoreAPI.Translate("info", "System is rebooting... Please wait a few minutes before reconnecting.")
+		}
 		w.Write([]byte(fmt.Sprintf(`<div class="alert alert-success">
 			<i class="bi bi-check-circle me-2"></i>
 			%s
